+++ modules/gui/skins2/vars/time.cpp
@@ -2,7 +2,7 @@
  * time.cpp
  *****************************************************************************
  * Copyright (C) 2003 VideoLAN
- * $Id: time.cpp,v 1.1 2004/01/03 23:31:34 asmax Exp $
+ * $Id: time.cpp,v 1.2 2004/01/11 17:12:17 asmax Exp $
  *
  * Authors: Cyril Deguet     <asmax@via.ecp.fr>
  *          Olivier Teulière <ipkiss@via.ecp.fr>
@@ -28,7 +28,7 @@
 #include <vlc/input.h>
 
 
-void Time::set( double percentage, bool updateVLC )
+void Time::set( float percentage, bool updateVLC )
 {
     if( getIntf()->p_sys->p_input == NULL )
     {
